use rated screw shaft clearance size
This is a reactionary change to something we should've done in the first
place. The countersink is very tight around the screw and makes
attaching the clamp very difficult. This should loosen that hold a bit,
and allow the screw to freely reach the nut where it can thread there.