switch to two-screw wall frame
By switching to two screws, we can ensure a purely vertical lineup with
the screws. Attaching with just two screws into drywall can sometimes be
tricky enough - using three where one was off to the side wasn't
helpful. This also removes the cross support at the recommendation of a
retired mechanical engineer (name unknown).

A hole in the support was added to make it easier to screw into, but
it's still proven difficult since the entire piece is just so narrow
compared to screws, screw bits, and screw drivers. I think instead of a
hole I should make the support very narrow, to allow tooling to come in
at a slight angle. This will come later though, and is not for this
commit.